test(gui): keep the desktop-rpc contract suite byte-for-byte stable
The synthesized-event mapping coverage (rtl2gds command mapping,
repeated-sequence tolerance) moves into its own file so the existing
contract suite remains an untouched regression net for AC-7.
